Create nested folders when saving movie stills
Since we have improved the async mode, the movies'
pages are visited very quickly even before the images
have started downloading.

To avoid having a lot of empty folders (until the scrapers
download the images), we only create the movie's folder
when saving an image, if it doesn't exist yet.